# backend/docs/swagger/openapi.yaml
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/v1/users/update/status":{"put":{"tags":["user-endpoint"],"operationId":"updateUserStatusByEmail","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApplicationUserDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApplicationUserResponse"}}}}}}},"/api/v1/location/{id}":{"get":{"tags":["location-endpoint"],"operationId":"findById_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json;charset=UTF-8":{"schema":{"$ref":"#/components/schemas/LocationDto"}}}}}},"put":{"tags":["location-endpoint"],"operationId":"update","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LocationDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json;charset=UTF-8":{"schema":{"$ref":"#/components/schemas/LocationDto"}}}}}},"delete":{"tags":["location-endpoint"],"operationId":"delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/v1/address/{id}":{"get":{"tags":["address-endpoint"],"operationId":"findById_3","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json;charset=UTF-8":{"schema":{"$ref":"#/components/schemas/AddressDto"}}}}}},"put":{"tags":["address-endpoint"],"operationId":"update_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddressDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json;charset=UTF-8":{"schema":{"$ref":"#/components/schemas/AddressDto"}}}}}},"delete":{"tags":["address-endpoint"],"operationId":"delete_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/v1/users/registration":{"post":{"tags":["user-endpoint"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreateRequest"}}},"required":true},"responses":{"201":{"description":"Created","content":{"application/json;charset=UTF-8":{"schema":{"type":"object"}}}}}}},"/api/v1/messages":{"get":{"tags":["message-endpoint"],"summary":"Get list of messages without details","operationId":"findAll","responses":{"200":{"description":"OK","content":{"application/json;charset=UTF-8":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SimpleMessageDto"}}}}}},"security":[{"apiKey":[]}]},"post":{"tags":["message-endpoint"],"summary":"Publish a new message","operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageInquiryDto"}}},"required":true},"responses":{"201":{"description":"Created","content":{"application/json;charset=UTF-8":{"schema":{"$ref":"#/components/schemas/DetailedMessageDto"}}}}},"security":[{"apiKey":[]}]}},"/api/v1/location":{"get":{"tags":["location-endpoint"],"operationId":"findAll_1","responses":{"200":{"description":"OK","content":{"application/json;charset=UTF-8":{"schema":{"type":"object"}}}}}},"post":{"tags":["location-endpoint"],"operationId":"create_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LocationCreateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json;charset=UTF-8":{"schema":{"$ref":"#/components/schemas/LocationDto"}}}}}}},"/api/v1/authentication":{"post":{"tags":["login-endpoint"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLoginDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json;charset=UTF-8":{"schema":{"type":"string"}}}}}}},"/api/v1/address":{"get":{"tags":["address-endpoint"],"operationId":"findAll_2","responses":{"200":{"description":"OK","content":{"application/json;charset=UTF-8":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AddressDto"}}}}}}},"post":{"tags":["address-endpoint"],"operationId":"create_2","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddressCreateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json;charset=UTF-8":{"schema":{"$ref":"#/components/schemas/AddressDto"}}}}}}},"/health":{"get":{"tags":["custom-health-endpoint"],"operationId":"getHealth","responses":{"200":{"description":"OK","content":{"application/json;charset=UTF-8":{"schema":{"type":"string"}}}}}}},"/health/prepareShutdown":{"get":{"tags":["custom-health-endpoint"],"operationId":"preShutdown","responses":{"200":{"description":"OK"}}}},"/api/v1/users/search":{"get":{"tags":["user-endpoint"],"operationId":"searchUsers","parameters":[{"name":"firstName","in":"query","required":false,"schema":{"type":"string"}},{"name":"familyName","in":"query","required":false,"schema":{"type":"string"}},{"name":"email","in":"query","required":false,"schema":{"type":"string"}},{"name":"isLocked","in":"query","required":false,"schema":{"type":"boolean"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ApplicationUserDto"}}}}}}}},"/api/v1/users/current":{"get":{"tags":["user-endpoint"],"operationId":"getUser","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApplicationUserResponse"}}}}}}},"/api/v1/tickets/ticket/{id}":{"get":{"tags":["ticket-endpoint"],"operationId":"findById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TicketDetailsResponse"}}}}}},"delete":{"tags":["ticket-endpoint"],"operationId":"cancelReservedTicket","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"204":{"description":"No Content"}}}},"/api/v1/tickets/mytickets":{"get":{"tags":["ticket-endpoint"],"operationId":"findForUser","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TicketDetailsResponse"}}}}}}}},"/api/v1/orders/order/{id}":{"get":{"tags":["order-endpoint"],"operationId":"findById_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderDetailsResponse"}}}}}},"delete":{"tags":["order-endpoint"],"operationId":"cancelOrder","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"204":{"description":"No Content"}}}},"/api/v1/orders/myorders":{"get":{"tags":["order-endpoint"],"operationId":"findForUser_1","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/OrderSummaryResponse"}}}}}}}},"/api/v1/messages/{id}":{"get":{"tags":["message-endpoint"],"summary":"Get detailed information about a specific message","operationId":"find","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json;charset=UTF-8":{"schema":{"$ref":"#/components/schemas/DetailedMessageDto"}}}}},"security":[{"apiKey":[]}]}},"/api/v1/location/search":{"get":{"tags":["location-endpoint"],"operationId":"search","parameters":[{"name":"name","in":"query","required":false,"schema":{"type":"string"}},{"name":"city","in":"query","required":false,"schema":{"type":"string"}},{"name":"street","in":"query","required":false,"schema":{"type":"string"}},{"name":"postalCode","in":"query","required":false,"schema":{"type":"string"}},{"name":"country","in":"query","required":false,"schema":{"type":"string"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":15}},{"name":"sort","in":"query","required":false,"schema":{"type":"string","default":"name"}}],"responses":{"200":{"description":"OK","content":{"application/json;charset=UTF-8":{"schema":{"$ref":"#/components/schemas/PageLocationDto"}}}}}}}},"components":{"schemas":{"AddressDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"street":{"type":"string"},"city":{"type":"string"},"zip":{"type":"string"},"country":{"type":"string"}}},"ApplicationUserDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"email":{"type":"string"},"password":{"type":"string"},"firstName":{"type":"string"},"familyName":{"type":"string"},"phoneNumber":{"type":"string"},"salt":{"type":"string"},"loginCount":{"type":"integer","format":"int32"},"accountLocked":{"type":"boolean"},"address":{"$ref":"#/components/schemas/AddressDto"},"admin":{"type":"boolean"},"superAdmin":{"type":"boolean"}}},"AddressResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"street":{"type":"string"},"city":{"type":"string"},"zip":{"type":"string"},"country":{"type":"string"}}},"ApplicationUserResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"email":{"type":"string"},"firstName":{"type":"string"},"familyName":{"type":"string"},"phoneNumber":{"type":"string"},"accountLocked":{"type":"boolean"},"isAdmin":{"type":"boolean"},"isSuperAdmin":{"type":"boolean"},"address":{"$ref":"#/components/schemas/AddressResponse"}}},"LocationDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"address":{"$ref":"#/components/schemas/AddressDto"}}},"AddressCreateRequest":{"type":"object","properties":{"street":{"type":"string"},"city":{"type":"string"},"zip":{"type":"string"},"country":{"type":"string"}}},"UserCreateRequest":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"},"firstName":{"type":"string"},"familyName":{"type":"string"},"phoneNumber":{"type":"string"},"isAdmin":{"type":"boolean"},"addressCreateRequest":{"$ref":"#/components/schemas/AddressCreateRequest"}}},"MessageInquiryDto":{"required":["summary","text","title"],"type":"object","properties":{"title":{"maxLength":100,"minLength":0,"type":"string"},"summary":{"maxLength":500,"minLength":0,"type":"string"},"text":{"maxLength":10000,"minLength":0,"type":"string"}}},"DetailedMessageDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"publishedAt":{"type":"string","format":"date-time"},"title":{"type":"string"},"summary":{"type":"string"},"text":{"type":"string"}}},"LocationCreateRequest":{"type":"object","properties":{"name":{"type":"string"},"addressCreateRequest":{"$ref":"#/components/schemas/AddressCreateRequest"}}},"UserLoginDto":{"required":["email","password"],"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}}},"EventResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"title":{"type":"string"},"description":{"type":"string"},"eventType":{"type":"string","enum":["CONCERT","THEATER","PLAY"]},"duration":{"type":"integer","format":"int64"}}},"HallSeatResponse":{"type":"object","description":"Hall-Seat","allOf":[{"$ref":"#/components/schemas/HallSpotResponse"},{"type":"object","properties":{"frontendCoordinates":{"type":"string"}}}]},"HallSectorResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"hallSectorShow":{"$ref":"#/components/schemas/HallSectorShowResponse"}}},"HallSectorShowResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"show":{"$ref":"#/components/schemas/ShowResponse"},"sector":{"$ref":"#/components/schemas/HallSectorResponse"},"price":{"type":"integer","format":"int64"}}},"HallSpotResponse":{"required":["type"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"sector":{"$ref":"#/components/schemas/HallSectorResponse"},"type":{"type":"string"}},"description":"Generic Hall-Spot","discriminator":{"propertyName":"type"}},"ShowResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"dateTime":{"type":"string","format":"date-time"},"event":{"$ref":"#/components/schemas/EventResponse"}}},"TicketDetailsResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"hash":{"type":"string"},"reserved":{"type":"boolean"},"valid":{"type":"boolean"},"hallSpot":{"oneOf":[{"$ref":"#/components/schemas/HallSpotResponse"},{"$ref":"#/components/schemas/HallSeatResponse"}]},"show":{"$ref":"#/components/schemas/ShowResponse"}}},"InvoiceResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"invoiceType":{"type":"string","enum":["PURCHASE","CANCELLATION"]},"dateTime":{"type":"string","format":"date-time"}}},"OrderDetailsResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"tickets":{"type":"array","items":{"$ref":"#/components/schemas/TicketDetailsResponse"}},"customer":{"$ref":"#/components/schemas/ApplicationUserResponse"},"invoices":{"type":"array","items":{"$ref":"#/components/schemas/InvoiceResponse"}},"dateTime":{"type":"string","format":"date-time"}}},"OrderSummaryResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"ticketCount":{"type":"integer","format":"int32"},"totalPrice":{"type":"integer","format":"int64"},"invoices":{"type":"array","items":{"$ref":"#/components/schemas/InvoiceResponse"}},"dateTime":{"type":"string","format":"date-time"}}},"SimpleMessageDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"publishedAt":{"type":"string","format":"date-time"},"title":{"type":"string"},"summary":{"type":"string"}}},"PageLocationDto":{"type":"object","properties":{"totalPages":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"first":{"type":"boolean"},"last":{"type":"boolean"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/LocationDto"}},"number":{"type":"integer","format":"int32"},"sort":{"type":"array","items":{"$ref":"#/components/schemas/SortObject"}},"numberOfElements":{"type":"integer","format":"int32"},"empty":{"type":"boolean"}}},"PageableObject":{"type":"object","properties":{"pageNumber":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"},"offset":{"type":"integer","format":"int64"},"sort":{"type":"array","items":{"$ref":"#/components/schemas/SortObject"}},"paged":{"type":"boolean"},"unpaged":{"type":"boolean"}}},"SortObject":{"type":"object","properties":{"direction":{"type":"string"},"nullHandling":{"type":"string"},"ascending":{"type":"boolean"},"property":{"type":"string"},"ignoreCase":{"type":"boolean"}}}}}}